try and remove the extra sstuff in the stream sent from veetle.
comparing the stream as sent from veetle and the stream between the two veetle vlc players shows that there is some extra stuff that is stripped out before sending on to the player.

this is a first attempt to find and remove this stuff. it doesn't work fully but some streams are detected by mplayer and sometimes a small bit is played.

somethin to investigate further.....